refactor: 컴포넌트 아키텍처 리팩토링
#66
