+++ src/components/Button/index.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import { ActivityIndicator, TouchableOpacityProps } from 'react-native';
-import styled, { css } from 'styled-components/native';
 import Text, { TextProps } from '@src/components/Text';
 import { useTheme } from '@src/contexts/ThemeProvider';
 import Icon, { IconProps } from '@src/components/Icon';
-import { flexCenter } from '@src/styles/flex';
+import { StyledButton } from './styled';
 
 type P = TouchableOpacityProps &
   TextProps & {
@@ -15,27 +14,11 @@ type P = TouchableOpacityProps &
     disabled?: boolean;
   };
 
-const RNTouchableOpacity = styled.TouchableOpacity<P>`
-  ${({ type = 'solid', theme, color }) =>
-    type === 'solid'
-      ? css`
-          background-color: ${theme[color || 'primary']};
-        `
-      : type === 'outline' &&
-        css`
-          border: 1px solid ${theme[color || 'primary']};
-        `}
-
-  opacity: ${({ disabled }) => (disabled ? 0.6 : 1)};
-  padding: ${({ type }) => (type === 'clear' ? 0 : '14px 20px')};
-  ${flexCenter}
-`;
-
 const Button: React.FC<P> = props => {
   const { colors } = useTheme();
 
   return (
-    <RNTouchableOpacity {...props}>
+    <StyledButton {...props}>
       {props.loading ? (
         <ActivityIndicator
           size={props.size || 16}
@@ -54,7 +37,7 @@ const Button: React.FC<P> = props => {
       ) : (
         props.icon && <Icon {...props.icon} />
       )}
-    </RNTouchableOpacity>
+    </StyledButton>
   );
 };
