+++ src/components/Loader/index.tsx
@@ -1,18 +1,7 @@
 import React from 'react';
 import { ActivityIndicator } from 'react-native';
-import styled from 'styled-components/native';
-import Text from '@src/components/Text';
-import { flexFillCenter } from '@src/styles/flex';
 import { useTheme } from '@src/contexts/ThemeProvider';
-
-const Container = styled.View`
-  background-color: ${props => props.theme.background};
-  ${flexFillCenter}
-`;
-
-const Title = styled(Text)`
-  margin-top: 10px;
-`;
+import { Container, Title } from './styled';
 
 const Loader: React.FC = () => {
   const { colors } = useTheme();
