+++ src/navigations/AuthNavigator.tsx
@@ -3,11 +3,11 @@ import {
   createNativeStackNavigator,
   NativeStackScreenProps,
 } from '@react-navigation/native-stack';
-import LoginScreen from '@src/screens/auth/LoginScreen';
-import RegisterScreen from '@src/screens/auth/RegisterScreen';
+import LoginScreen from '@src/screens/LoginScreen';
+import RegisterScreen from '@src/screens/RegisterScreen';
 import { RootStackParamList } from '@src/types/navigation';
 import { StackNavigatorDefaultScreenOptions } from '@src/navigations/StackNavigatorDefaultScreenOptions';
-import SuccessModalScreen from '@src/screens/auth/SuccessModalScreen';
+import SuccessModalScreen from '@src/screens/SuccessModalScreen';
 
 const Stack = createNativeStackNavigator();
