+++ src/screens/HomeScreen/index.tsx
@@ -3,14 +3,9 @@ import { MainTabParamList, RootStackParamList } from '@src/types/navigation';
 import { BottomTabScreenProps } from '@react-navigation/bottom-tabs';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
 import { CompositeScreenProps } from '@react-navigation/native';
-import styled from 'styled-components/native';
-import { flexFillCenter } from '@src/styles/flex';
 import Text from '@src/components/Text';
 import Button from '@src/components/Button';
-
-const Container = styled.View`
-  ${flexFillCenter}
-`;
+import { Container } from './styled';
 
 type P = CompositeScreenProps<
   BottomTabScreenProps<MainTabParamList, 'Home'>,
