+++ src/screens/Profile/styled.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components/native';
+import { flexFillCenter } from '@src/styles/flex';
+
+export const Container = styled.View`
+  ${flexFillCenter}
+`;
