+++ src/screens/RegisterScreen/index.tsx
@@ -2,13 +2,8 @@ import React from 'react';
 import { CompositeScreenProps } from '@react-navigation/native';
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
 import { AuthStackParamList, RootStackParamList } from '@src/types/navigation';
-import styled from 'styled-components/native';
-import { flexFillCenter } from '@src/styles/flex';
 import Text from '@src/components/Text';
-
-const Container = styled.View`
-  ${flexFillCenter}
-`;
+import { Container } from './styled';
 
 type P = CompositeScreenProps<
   NativeStackScreenProps<AuthStackParamList, 'Login'>,
